fixed name collision when Quest ID is equal to username
When a Quest ID is the same as the system username the Quest password
concatenated with the username was treated as the default user. The
mitigation for this was to prefix the current username with "_config_"
and treat that as the username to store the user comma separated list.
To make things more clear the variable name 'current' was changed to
'key_to_users_csv'.

This fix was manually tested. It does not seem like this project has
unit tests although a harness exists.